delete post

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -41,31 +41,24 @@ const Mutation = {
       token: jwt.sign({ userId: user.id }, "thisIsASecret")
     };
   },
-  async deleteUser(parent, args, { prisma }, info) {
-    const userExists = await prisma.exists.User({ id: args.id });
-    if (!userExists) {
-      throw new Error("User not found");
-    }
-    return prisma.mutation.deleteUser({ where: { id: args.id } }, info);
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    return prisma.mutation.deleteUser({ where: { id: userId } }, info);
   },
-  updateUser(parent, args, { prisma }, info) {
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.updateUser(
       {
         where: {
-          id: args.id
+          id: userId
         },
         data: args.data
       },
       info
     );
   },
-  async createPost(parent, args, { prisma, request }, info) {
+  createPost(parent, args, { prisma, request }, info) {
     const userId = getUserId(request);
-    const userExists = await prisma.exists.User({ id: args.id });
-    if (!userExists) {
-      throw new Error("User not found");
-    }
-    console.log(userId);
     return prisma.mutation.createPost(
       {
         data: {
@@ -82,7 +75,17 @@ const Mutation = {
       info
     );
   },
-  deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if (!postExists) {
+      throw new Error("Unable to delete post");
+    }
     return prisma.mutation.deletePost({ where: { id: args.id } }, info);
   },
   updatePost(parent, args, { prisma }, info) {
